pyproject.toml
[tool.poetry]
name = "backend-product"
version = "0.1.0"
description = ""
authors = ["Daniel Alejandro Leyva Perez <dalejandrolp99@gmail.com>"]
readme = "README.md"
packages = [{ include = "backend_product" }]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.7"
whitenoise = {extras = ["brotli"], version = "^6.5.0"}
gunicorn = "^21.2.0"
graphene-django = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
